show product page when funds loaded for reservation block
