+++ react/src/webshop/components/pages/products-show/ProductsShow.tsx
@@ -24,6 +24,8 @@ import useEnvData from '../../../hooks/useEnvData';
 import BlockShowcase from '../../elements/block-showcase/BlockShowcase';
 import useSetProgress from '../../../../dashboard/hooks/useSetProgress';
 import { clickOnKeyEnter } from '../../../../dashboard/helpers/wcag';
+import Fund from '../../../props/models/Fund';
+import { useFundService } from '../../../services/FundService';
 
 export default function ProductsShow() {
     const { id } = useParams();
@@ -39,10 +41,12 @@ export default function ProductsShow() {
     const navigateState = useNavigateState();
     const bookmarkProductToggle = useBookmarkProductToggle();
 
+    const fundService = useFundService();
     const productService = useProductService();
     const voucherService = useVoucherService();
     const providerService = useProviderService();
 
+    const [funds, setFunds] = useState<Array<Fund>>(null);
     const [product, setProduct] = useState<Product>(null);
     const [provider, setProvider] = useState<Provider>(null);
     const [vouchers, setVouchers] = useState<Array<Voucher>>(null);
@@ -101,6 +105,20 @@ export default function ProductsShow() {
             .then((res) => setVouchers(res.data.data))
             .finally(() => setProgress(100));
     }, [authIdentity, voucherService, setProgress, id]);
+
+    const fetchFunds = useCallback(() => {
+        if (!product?.funds.length) {
+            return;
+        }
+
+        fundService
+            .list({ per_page: 100, check_criteria: 1, fund_ids: product?.funds.map((fund) => fund.id) })
+            .then((res) => setFunds(res.data.data));
+    }, [fundService, product?.funds]);
+
+    useEffect(() => {
+        fetchFunds();
+    }, [fetchFunds]);
 
     useEffect(() => {
         fetchProduct();
@@ -170,7 +188,7 @@ export default function ProductsShow() {
                     </div>
                 )
             }>
-            {product && (
+            {product && funds && vouchers && (
                 <section className="section section-product">
                     <div className="block block-product">
                         <div className="product-card">
@@ -232,7 +250,7 @@ export default function ProductsShow() {
                             )}
                         </div>
 
-                        {product && vouchers && <ProductFundsCard product={product} vouchers={vouchers} />}
+                        <ProductFundsCard product={product} vouchers={vouchers} funds={funds} />
 
                         {provider && (
                             <div className="block block-organizations">
